Show decorations before messages loaded, and ensure update after edit
Befoer it waited for all `git show` commands to finish before setting the
decoration. Now it sets it right away, and then again after they are done,
unless something else triggered a new update in the meantime.

I was seeing the decoration move to the start of the line when deleting a
line with Cmd+Shift+K. I tried all 4 options of Closed/Open and it didn't
help, so just added an updateEditor call in the edit callback. Possibly
wasteful since most of the time there will also be a selection change.
Could add back the 'last selection' thing, or debounce.