Attach decorations to end of line (Fixes #4)
Originally I attached decorations to the empty range at the end of the
line with isWholeLine=false. In 3284b6d7b7ac6cf9ebe37e707afa4c852fb7224c
I changed it to use the entire line's range with isWholeLine=true. I am
now reverting that, and going back to the original approach.

This ensures that the hover message only shows up when you actually
hover over the blame decoration, rather than hovering anywhere on the
line. The latter is undesirable because then it gets confusingly
combined with other hover messages, e.g. types and doc comments.

There is one downside: if you join lines (e.g. with Ctrl-J or deleting
forward instead of backward), the inline annotation can sometimes show
briefly in the middle of the line before it gets updated. With
isWholeLine VS Code knows to always put it at the end. I think this used
to be a worse problem because I had less precise updating, so perhaps it
showed in the middle until something finished loading. Anyway, when I
test Ctrl-J now, most of the time I can't even see a single frame where
it's in the middle. But if I do it a bunch of times in a row, I can
sometimes see the tiniest flicker.

So I think the pros outweigh the cons. If anyone really has problems
with the flicker then I can add a config to restore the old behavior.