Add config betterGitLineBlame.annotateWholeLine
In ed92fe5082be082d9f019199895bae867db1d41a I was pretty convinced that
the annotation flickering is barely noticeable, but I forgot about
remote development. There, it is much more noticeable, so I think it's
worth having a setting to go back to the old behavior and avoid flicker,
at the cost of having the hover message show up when you hover anywhere
on the line.