Ensure annotation is always at the end of the line
Without isWholeLine=true, the annotation can end up in the middle of the
line if updating decorations races with typing. By setting the column to
Number.MAX_VALUE, it always goes at the end of the line. This is not a
documented feature as far as I know, it simply clamps to the actual end
of the line if the column is too large.