Avoid removing status bar item too often
We remove it if you click into an editor that isn't in git or isn't even
a normal file, like a log window. But if you just have that window open,
every log that gets printed causes the selection change event. So now we
check if it's actually the active editor.