Avoid excessive decoration updating
Add another level of caching so that we don't try to update decorations
if it's the same line and nothing has changed. All updateEditor() calls
by the extension (as opposed to the user changing text selection) pass
{ force: true } because I don't trust myself to get the caching logic
100% right. But I've tested it for a while and it seems to be working.

I had to split up the first part of getCommitInfo into getCommitTemp
(couldn't think of a better name) so that we just get the state first
avoid loading more info until we're sure we need to update.